move to common place
