+++ src/queries/TrackingGQL.js
@@ -1,0 +1,7 @@
+import gql from 'graphql-tag'
+
+export const TRACK_EVENTS_MUTATION = gql`
+  mutation trackEvents($events: json) {
+    trackEvents(events: $events)
+  }
+`
